Added HttpEmail, reworked extract.py, added trec-cache
HttpEmail was required to be able to easily read in a single
email file when doing the training/testing.

Reworking extract.py using HttpEmail to provide the code more
clarity.

Added trec-cache as an organized collection of word extracts
for each file and the whole repository as a whole.

<numeric-id>.<spam|ham>.table		files containing the
					word tables for the
					associated email file.

TABLE.spam				complete word table for
					the spam dataset.

TABLE.ham				complete word table for
					the ham dataset.